src/app/backend/requirements.txt
fastapi
uvicorn
torch
torchvision
matplotlib
tqdm
streamlit
opencv-python-headless
pillow
requests
streamlit-drawable-canvas